+++ packages/bluesky/lib/bluesky.dart
@@ -302,10 +302,10 @@ export 'package:bluesky/src/services/entities/skeleton_post.dart';
 export 'package:bluesky/src/services/entities/skeleton_actors_by_query.dart';
 export 'package:bluesky/src/services/entities/skeleton_actor.dart';
 export 'package:bluesky/src/services/entities/suggested_follows.dart';
-export 'package:bluesky/src/services/extensions/grouped_notification_extension.dart';
+export 'package:bluesky/src/services/extensions/grouped_notification.dart';
 export 'package:bluesky/src/services/extensions/at_uri_extension.dart';
-export 'package:bluesky/src/services/extensions/blob_extension.dart';
-export 'package:bluesky/src/services/extensions/strong_ref_extension.dart';
+export 'package:bluesky/src/services/extensions/blob.dart';
+export 'package:bluesky/src/services/extensions/strong_ref.dart';
 export 'package:bluesky/src/services/feed_service.dart';
 export 'package:bluesky/src/services/extensions/feed_service.dart';
 export 'package:bluesky/src/services/graph_service.dart';
