+++ packages/bluesky/lib/src/services/extensions/strong_ref.dart
