+++ packages/bluesky/test/src/services/extensions/blob_extension_test.dart
@@ -9,7 +9,7 @@ import 'package:test/test.dart';
 // 🌎 Project imports:
 import 'package:bluesky/src/services/entities/embed.dart';
 import 'package:bluesky/src/services/entities/image.dart';
-import 'package:bluesky/src/services/extensions/blob_extension.dart';
+import 'package:bluesky/src/services/extensions/blob.dart';
 
 void main() {
   test('.toImage', () {
