+++ packages/bluesky/test/src/services/extensions/strong_ref_extension_test.dart
@@ -8,7 +8,7 @@ import 'package:test/test.dart';
 
 // 🌎 Project imports:
 import 'package:bluesky/src/services/entities/embed_record.dart';
-import 'package:bluesky/src/services/extensions/strong_ref_extension.dart';
+import 'package:bluesky/src/services/extensions/strong_ref.dart';
 
 void main() {
   test('.toEmbedRecord', () {
